move shared capri calculations to libcapri
